requirements.txt
pyodbc
shillelagh
gsheetsdb
gspread
oauth2client
configparser
streamlit_authenticator
opencv-python